Finish pdef API calls

+++ dqcsim/src/plugin/definition.rs
@@ -13,7 +13,8 @@ use std::fmt;
 ///
 /// This struct is constructed by the user (or the foreign language API). The
 /// behavior of the plugin is defined by a number of closures that the user
-/// must provide.
+/// must provide. For more information about the callback functions, refer to
+/// the documentation of the foreign language API setter functions.
 #[allow(clippy::type_complexity)]
 pub struct PluginDefinition {
     /// Plugin type.
@@ -23,126 +24,26 @@ pub struct PluginDefinition {
     metadata: PluginMetadata,
 
     /// Initialization callback.
-    ///
-    /// This is always called before any of the other callbacks are run. The
-    /// downstream plugin has already been initialized at this stage, so it is
-    /// legal to send it commands.
-    ///
-    /// The default behavior is no-op.
     pub initialize: Box<dyn Fn(&mut PluginState, Vec<ArbCmd>) -> Result<()> + Send + 'static>,
 
     /// Finalization callback.
-    ///
-    /// This is called when a plugin is gracefully terminated.
-    ///
-    /// The default behavior is no-op.
     pub drop: Box<dyn Fn(&mut PluginState) -> Result<()> + Send + 'static>,
 
     /// Run callback for frontends.
-    ///
-    /// This is called in response to a `start()` host API call. The return
-    /// value is returned through the `wait()` host API call.
-    ///
-    /// The default behavior is to fail with a "not implemented" error;
-    /// frontends backends should always override this. This callback is never
-    /// called for operator or backend plugins.
     pub run: Box<dyn Fn(&mut PluginState, ArbData) -> Result<ArbData> + Send + 'static>,
 
     /// Qubit allocation callback for operators and backends.
-    ///
-    /// The default for operators is to pass through to `state.allocate()`. The
-    /// default for backends is no-op. This callback is never called for
-    /// frontend plugins.
     pub allocate:
         Box<dyn Fn(&mut PluginState, Vec<QubitRef>, Vec<ArbCmd>) -> Result<()> + Send + 'static>,
 
     /// Qubit deallocation callback for operators and backends.
-    ///
-    /// The default for operators is to pass through to `state.free()`. The
-    /// default for backends is no-op. This callback is never called for
-    /// frontend plugins.
     pub free: Box<dyn Fn(&mut PluginState, Vec<QubitRef>) -> Result<()> + Send + 'static>,
 
     /// Gate execution callback for operators and backends.
-    ///
-    /// Backends must return any measured qubits through the return value. Only
-    /// qubits that are part of the `measures` vector may be returned in this
-    /// vector.
-    ///
-    /// The story is more complicated for operators. Let's say we want to make
-    /// a silly operator that inverts all measurements. While the trivial way
-    /// to do this would be something like:
-    ///
-    /// ```ignore
-    /// fn gate(...) -> ... {
-    ///     state.gate(...)?;
-    ///     let mut measurements;
-    ///     for qubit in measures {
-    ///         measurements.push(QubitMeasurementResult {
-    ///             qubit,
-    ///             value: !state.get_measurement(qubit)?, // note the !
-    ///             data: state.get_measurement_arb(qubit)?,
-    ///         });
-    ///     }
-    ///     Ok(measurements)
-    /// }
-    /// ```
-    ///
-    /// this pattern should be avoided, because `get_measurement()` prevents
-    /// the operator thread from continuing until the downstream plugin has
-    /// returned the measurement, which may slow down the simulation. Instead,
-    /// you can return an empty vector (or leave the function unimplemented as
-    /// this is its default behavior) and use `modify_measurements()` for this:
-    ///
-    /// ```ignore
-    /// fn gate(...) -> ... {
-    ///     state.gate(...)?;
-    /// }
-    ///
-    /// fn modify_measurement(...) -> ... {
-    ///     measurement.value = !measurement.value;
-    ///     Ok(vec![measurement])
-    /// }
-    /// ```
-    ///
-    /// `gate()` is called when we get the request to execute a gate from
-    /// upstream, while `modify_measurement()` is called when we receive the
-    /// result from downstream.
-    ///
-    /// For more complex operators this method may become unwieldly. It's still
-    /// fine to implement everything in `gate()` in this case. However, if you
-    /// do this, override `modify_measurement()` to always return an empty
-    /// vector, otherwise the two methods may interfere with each other!
-    ///
-    /// The default behavior for operators is to pass through to `state.gate()`.
-    /// An empty measurements vector.is returned in this case; the measurements
-    /// are instead propagated through the `modify_measurements()` callback.
-    /// The default for backends is to fail with a "not implemented" error;
-    /// backends should always override this. This callback is never called for
-    /// frontend plugins.
     pub gate:
         Box<dyn Fn(&mut PluginState, Gate) -> Result<Vec<QubitMeasurementResult>> + Send + 'static>,
 
     /// Measurement modification callback for operators.
-    ///
-    /// This callback is somewhat special, in that it is not allowed to call
-    /// any plugin command other than logging and the pseudorandom number
-    /// generator functions. This is because this function is called
-    /// asynchronously with respect to the downstream functions, making the
-    /// timing of these calls non-deterministic based on operating system
-    /// scheduling.
-    ///
-    /// Note that while this function is called for only a single measurement
-    /// at a time, it is allowed to produce a vector of measurements. This
-    /// allows you to cancel propagation of the measurement by returning an
-    /// empty vector, to just modify the measurement data itself, or to
-    /// generate additional measurements from a single measurement. However,
-    /// if you need to modify the qubit references for operators that remap
-    /// qubits, take care to only send measurement data upstream when this is
-    /// explicitly requested through a gate function's `measured` list.
-    ///
-    /// The default behavior for this callback is to return the measurement
-    /// without modification.
     pub modify_measurement: Box<
         dyn Fn(&mut PluginState, QubitMeasurementResult) -> Result<Vec<QubitMeasurementResult>>
             + Send
@@ -150,24 +51,13 @@ pub struct PluginDefinition {
     >,
 
     /// Callback for advancing time for operators and backends.
-    ///
-    /// The default behavior for operators is to pass through to
-    /// `state.advance()`. The default for backends is no-op. This callback is
-    /// never called for frontend plugins.
-    pub advance: Box<dyn Fn(&mut PluginState, usize) -> Result<()> + Send + 'static>,
+    pub advance: Box<dyn Fn(&mut PluginState, u64) -> Result<()> + Send + 'static>,
 
     /// Callback function for handling an arb from upstream for operators and
     /// backends.
-    ///
-    /// The default behavior for operators is to pass through to
-    /// `state.arb()`; operators that do not support the requested interface
-    /// should always do this. The default for backends is no-op. This callback
-    /// is never called for frontend plugins.
     pub upstream_arb: Box<dyn Fn(&mut PluginState, ArbCmd) -> Result<ArbData> + Send + 'static>,
 
     /// Callback function for handling an arb from the host.
-    ///
-    /// The default behavior for this is no-op.
     pub host_arb: Box<dyn Fn(&mut PluginState, ArbCmd) -> Result<ArbData> + Send + 'static>,
 }
